Rabin-Karp and some style improvements
The search now uses the Rabin-Karp algorithm.
The patterns now need to be provided in a text file with the following structure:

int Priority; String "Pattern"; String "Result"
int Priority; String "Pattern"; String "Result"
...

Added a pattern object to hold the patterns from the file.

Removed the file extractor. Extracting of the files from the folder is now happening as a stream in the main function.

Added some lambdas for short functions to improve readability.

Updated the topics.

Finished working on this project